+++ src/tabmat/__init__.py
@@ -1,7 +1,7 @@
 import importlib.metadata
 
 from .categorical_matrix import CategoricalMatrix
-from .constructor import from_csc, from_formula, from_pandas
+from .constructor import from_csc, from_formula, from_pandas, from_polars
 from .dense_matrix import DenseMatrix
 from .matrix_base import MatrixBase
 from .sparse_matrix import SparseMatrix
@@ -23,6 +23,7 @@ __all__ = [
     "from_csc",
     "from_formula",
     "from_pandas",
+    "from_polars",
     "as_tabmat",
     "hstack",
 ]